fastapi-app/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
blinker==1.9.0
cffi==1.17.1
click==8.2.1
colorama==0.4.6
cryptography==44.0.2
fastapi==0.115.12
flasgger==0.9.7.1
Flask==3.1.1
Flask-HTTPAuth==4.8.0
Flask-JWT-Extended==4.7.1
h11==0.16.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
MarkupSafe==3.0.2
mistune==3.1.3
packaging==25.0
pycparser==2.22
pycryptodome==3.22.0
pydantic==2.11.5
pydantic_core==2.33.2
PyJWT==2.10.1
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.25.1
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.13.2
uvicorn==0.34.3
Werkzeug==3.1.3
fastapi[all]
python-dotenv
python-jose[cryptography]
passlib[bcrypt]